requirements.txt
flask>=2.0.0
web3>=5.24.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
tensorflow>=2.8.0
discord.py>=1.7.3
python-dotenv>=0.19.0 